backend/requirements.txt
Django==5.2.13
djangorestframework==3.16.1
django-filter==25.2
django-cors-headers==4.9.0
dj-database-url==3.0.1
psycopg2-binary==2.9.11
python-dotenv==1.1.1
Pillow==11.3.0
djangorestframework-simplejwt==5.5.1
